requirements.txt
# Real-time Screen Translator - Dependencies

# GUI Framework (tkinter is included with Python, no need to install)
# Screen Capture
mss>=9.0.1
Pillow>=10.0.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0

# OCR - CPU-only mode (optimal for real-time gaming)
pytesseract>=0.3.10
easyocr>=1.7.0
# Note: easyocr automatically installs torch (CPU version)
# CPU-only mode provides better real-time performance for gaming translations

# Translation
deep-translator>=1.11.4
deepl>=1.12.0

# Optional: Encoding detection for cache files
chardet>=5.0.0

# Advanced image hashing for better deduplication (perceptual hash)
imagehash>=4.3.1

# Global hotkeys support
pynput>=1.7.6

